+++ CHANGELOG.md
@@ -1,1 +1,2 @@
 - Updated Simplified Chinese localization (#315)
+- fix restock amount also displaying in stacks in JEI (#318)